fix: fix lsp code warning

Created-by: languager_server_ty
Commit-by: languager_server_ty
Merged-by: TimZhou
Description: ## 变更内容（必填）

fix:fix lsp code warning

## 变更类型（必填）

请描述本次Pull Request变更类型（原因），**请在保存后点击复选框或在编辑时将对应项前面的 `[ ]` 改为 `[X]` 。**

- [ ] 新增需求
- [x] 问题修复
- [ ] 构建过程或辅助工具变动
- [ ] 文档更新

## 变更内容自检（必填）

**请勿修改或删除以下选项内容，请在保存后点击复选框或在编辑时将对应项前面的 `[ ]` 改为 `[X]` 。**

### 编译本地自验证结果：

- [x] 仓颉工具编译通过
- [ ] 不涉及，辅助工具变动或者文档更新可以直接选择此项

### 测试用例本地自验证结果：

- [ ] 通过，请在下方提供截图证明
- [x] 不涉及，请在下方说明情况

## 关联的issue（必填）

请填写此Pull Request要关联的issue链接。

See merge request: Cangjie/cangjie_tools!266

+++ cangjie-language-server/src/languageserver/index/IndexDatabase.cpp
@@ -1583,14 +1583,14 @@ dberr_no OpenIndexDatabase(IndexDatabase &db, const std::string &file,
     return db.Initialize([dbfile = file, dbOpts = std::move(opts)] {
         if (dbOpts.openInMemory) {
             return (sqldb::memdb)(
-                dbfile, sqldb::OpenURI | (dbOpts.openReadOnly ? sqldb::OpenReadOnly
-                                                          : (sqldb::OpenCreate |
-                                                             sqldb::OpenReadWrite)));
+                dbfile, (unsigned int)sqldb::OpenURI | (dbOpts.openReadOnly ? (unsigned int)sqldb::OpenReadOnly
+                                                          : ((unsigned int)sqldb::OpenCreate |
+                                                             (unsigned int)sqldb::OpenReadWrite)));
         } else {
             return (sqldb::open)(
-                dbfile, sqldb::OpenURI | (dbOpts.openReadOnly ? sqldb::OpenReadOnly
-                                                        : (sqldb::OpenCreate |
-                                                           sqldb::OpenReadWrite)));
+                dbfile, (unsigned int)sqldb::OpenURI | (dbOpts.openReadOnly ? (unsigned int)sqldb::OpenReadOnly
+                                                        : ((unsigned int)sqldb::OpenCreate |
+                                                           (unsigned int)sqldb::OpenReadWrite)));
         }
     });
 }
